also check video-url, before attempting to delete a video in the filesystem

+++ src/Bo.php
@@ -815,7 +815,7 @@ class Bo
 					count($comments), $answers));
 			}
 		}
-		if (!empty($video['video_hash']))
+		if (!empty($video['video_hash']) && empty($video['video_url']))
 		{
 			unlink($this->videoPath($video));
 		}
